Refactor TaskListTest
