+++ src/test/java/retriever/task/TaskListTest.java
@@ -4,11 +4,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-import java.util.ArrayList;
-
 import org.junit.jupiter.api.Test;
 
-import retriever.Storage;
+import retriever.StorageStub;
 import retriever.exception.IllegalDateFormatException;
 import retriever.exception.IllegalDeadlineFormatException;
 import retriever.exception.RetrieverException;
@@ -16,7 +14,7 @@ import retriever.exception.RetrieverException;
 public class TaskListTest {
     @Test
     public void taskListLength_checkingLengthOfEmptyList_success() {
-        assertEquals(0, new TaskList(new StorageStub()).taskListLength());
+        assertEquals(0, new TaskList(new StorageStub("file_name_stub")).taskListLength());
     }
 
     @Test
@@ -25,7 +23,7 @@ public class TaskListTest {
         String task = "deadline CS2103T iP /by 26/08/2021";
 
         try {
-            TaskList taskList = new TaskList(new StorageStub());
+            TaskList taskList = new TaskList(new StorageStub("file_name_stub"));
             taskList.addDeadlineTask(task);
             hasPassed = true;
         } catch (RetrieverException e) {
@@ -41,7 +39,7 @@ public class TaskListTest {
         String task = "event Apple WWDC 2021 /at 10/06/2021";
 
         try {
-            TaskList taskList = new TaskList(new StorageStub());
+            TaskList taskList = new TaskList(new StorageStub("file_name_stub"));
             taskList.addEventTask(task);
             hasPassed = true;
         } catch (RetrieverException e) {
@@ -57,7 +55,7 @@ public class TaskListTest {
         String task = "todo laundry";
 
         try {
-            TaskList taskList = new TaskList(new StorageStub());
+            TaskList taskList = new TaskList(new StorageStub("file_name_stub"));
             taskList.addTodoTask(task);
             hasPassed = true;
         } catch (RetrieverException e) {
@@ -69,7 +67,7 @@ public class TaskListTest {
 
     @Test
     public void taskListLength_afterAddingSomeItems_success() throws RetrieverException {
-        TaskList taskList = new TaskList(new StorageStub());
+        TaskList taskList = new TaskList(new StorageStub("file_name_stub"));
         taskList.addTodoTask("todo sleep");
         taskList.addDeadlineTask("deadline CS2103T iP /by 26/08/2021");
         assertEquals(2, taskList.taskListLength());
@@ -81,7 +79,7 @@ public class TaskListTest {
         String task = "deadline CS2103T iP /by 26-08-2021";
 
         try {
-            TaskList taskList = new TaskList(new StorageStub());
+            TaskList taskList = new TaskList(new StorageStub("file_name_stub"));
             taskList.addDeadlineTask(task);
             fail();
         } catch (IllegalDateFormatException | IllegalDeadlineFormatException e) {
@@ -96,7 +94,7 @@ public class TaskListTest {
         String task = "event Apple WWDC 2021 /at 10th June 2021";
 
         try {
-            TaskList taskList = new TaskList(new StorageStub());
+            TaskList taskList = new TaskList(new StorageStub("file_name_stub"));
             taskList.addEventTask(task);
             fail();
         } catch (RetrieverException e) {
@@ -106,33 +104,3 @@ public class TaskListTest {
     }
 }
 
-class StorageStub extends Storage {
-    public StorageStub() {
-        super("file_name_stub");
-    }
-
-    @Override
-    public ArrayList<Task> readTasks() {
-        return new ArrayList<Task>();
-    }
-
-    @Override
-    public void writeTask(Task.TaskType task, String taskDescription, String taskDeadline) {
-        // Assume we are writing a task to the file
-    }
-
-    @Override
-    public void deleteTask(int taskNumber) {
-        // Assume we are deleting task from file
-    }
-
-    @Override
-    public void updateTaskStatusToDone(int taskNumber) {
-        // Assume we are updating task status in file
-    }
-
-    @Override
-    public void createNewFile() {
-        // Assume we are creating a new file
-    }
-}
